fix: subscription downgrade issue and messages updated
